requirements.txt
flask
flask-cors
numpy
pillow
tensorflow